feat: useDraftLoader hook
